+++ app/scripts/options_controller.js
@@ -2,52 +2,25 @@
 
 ngApp.
   factory('mblistData', ['$q',function ($q) {
-    var _cache = {
-      'mbList':[]
-    },
-    needCheck = true
-    ;
-
-    function _getAll () {
-      var deferred = $q.defer();
-
-      chrome.storage.local.get('bmIdArr', function(arrData) {
-        _cache.mbList = [];
-        if(arrData.bmIdArr && arrData.bmIdArr.length) {
-          _cache.mbList = arrData.bmIdArr;
-          deferred.resolve(arrData.bmIdArr);
-          needCheck = false;
-        }else {
-          deferred.reject([]);
-        }
-      });
-
-      return deferred.promise;
-
-    }
-  
     return {
       /**
        * 获取所有list信息
        * @return 
        */
       getAll: function(){
-        if(needCheck || _cache.mbList.length === 0) {
-          return _getAll();
-        }else {
-          var deferred = $q.defer();
-          deferred.resolve(_cache.mbList);
-          return deferred.promise;
-        }
+        var deferred = $q.defer();
+        chrome.runtime.sendMessage({'action':'getAll'},function(data){
+          // console.log(data);
+          deferred.resolve(data);
+        });
+
+        return deferred.promise;
       },
       /**
        * 获取单个信息
        */
       getId: function (id) {
 
-      },
-      checkNext: function() {
-        needCheck = true;
       }
     };
   }])