+++ HISTORY.md
@@ -10,8 +10,15 @@
 
 ## 数据存储初步使用 20151017
 1. chrome.storage
+2. 除background 外其余页面均使用 chrome.runtime.sendMessage 管理数据
 
-## omnibox 基本设置
+## omnibox 基本设置 20151018
+1. keyword cbm
+2. 提示基础
+
+## 访问记录功能 20151018
+1. 所有http和http请求
+2. 通过content_scripts 统一通过 sendMessage 管理